refactor compose builder
