+++ src/docker/composebuilder/service_builder.go
@@ -11,8 +11,8 @@ import (
 	"github.com/Originate/exosphere/src/util"
 )
 
-// DevelopmentDockerComposeBuilder contains the docker-compose.yml config for a single service
-type DevelopmentDockerComposeBuilder struct {
+// ServiceComposeBuilder contains the docker-compose.yml config for a single service
+type ServiceComposeBuilder struct {
 	AppConfig                types.AppConfig
 	ServiceConfig            types.ServiceConfig
 	ServiceData              types.ServiceData
@@ -24,9 +24,17 @@ type DevelopmentDockerComposeBuilder struct {
 	ServiceEndpoints         map[string]*ServiceEndpoints
 }
 
-// NewDevelopmentDockerComposeBuilder is DevelopmentDockerComposeBuilder's constructor
-func NewDevelopmentDockerComposeBuilder(appConfig types.AppConfig, serviceConfig types.ServiceConfig, serviceData types.ServiceData, role, appDir string, mode BuildMode, serviceEndpoints map[string]*ServiceEndpoints) *DevelopmentDockerComposeBuilder {
-	return &DevelopmentDockerComposeBuilder{
+// GetServiceDockerConfig returns the DockerConfigs for a service and its dependencies in docker-compose.yml
+func GetServiceDockerConfig(appConfig types.AppConfig, serviceConfig types.ServiceConfig, serviceData types.ServiceData, role string, appDir string, mode BuildMode, serviceEndpoints map[string]*ServiceEndpoints) (types.DockerConfigs, error) {
+	return NewServiceComposeBuilder(appConfig, serviceConfig, serviceData, role, appDir, mode, serviceEndpoints).getServiceDockerConfigs()
+}
+
+// NewServiceComposeBuilder is ServiceComposeBuilder's constructor
+func NewServiceComposeBuilder(appConfig types.AppConfig, serviceConfig types.ServiceConfig, serviceData types.ServiceData, role, appDir string, mode BuildMode, serviceEndpoints map[string]*ServiceEndpoints) *ServiceComposeBuilder {
+	if mode.Environment == BuildModeEnvironmentTest {
+		role = appConfig.GetTestRole(role)
+	}
+	return &ServiceComposeBuilder{
 		AppConfig:                appConfig,
 		ServiceConfig:            serviceConfig,
 		ServiceData:              serviceData,
@@ -40,7 +48,7 @@ func NewDevelopmentDockerComposeBuilder(appConfig types.AppConfig, serviceConfig
 }
 
 // getServiceDockerConfigs returns a DockerConfig object for a single service and its dependencies (if any(
-func (d *DevelopmentDockerComposeBuilder) getServiceDockerConfigs() (types.DockerConfigs, error) {
+func (d *ServiceComposeBuilder) getServiceDockerConfigs() (types.DockerConfigs, error) {
 	if d.ServiceData.Location != "" {
 		return d.getInternalServiceDockerConfigs()
 	}
@@ -50,14 +58,14 @@ func (d *DevelopmentDockerComposeBuilder) getServiceDockerConfigs() (types.Docke
 	return types.DockerConfigs{}, fmt.Errorf("No location or docker image listed for '%s'", d.Role)
 }
 
-func (d *DevelopmentDockerComposeBuilder) getDockerfileName() string {
+func (d *ServiceComposeBuilder) getDockerfileName() string {
 	if d.Mode.Environment == BuildModeEnvironmentProduction {
 		return "Dockerfile.prod"
 	}
 	return "Dockerfile.dev"
 }
 
-func (d *DevelopmentDockerComposeBuilder) getDockerCommand() string {
+func (d *ServiceComposeBuilder) getDockerCommand() string {
 	switch d.Mode.Environment {
 	case BuildModeEnvironmentProduction:
 		return ""
@@ -68,7 +76,7 @@ func (d *DevelopmentDockerComposeBuilder) getDockerCommand() string {
 	}
 }
 
-func (d *DevelopmentDockerComposeBuilder) getDockerPorts() []string {
+func (d *ServiceComposeBuilder) getDockerPorts() []string {
 	switch d.Mode.Environment {
 	case BuildModeEnvironmentProduction:
 		fallthrough
@@ -79,21 +87,21 @@ func (d *DevelopmentDockerComposeBuilder) getDockerPorts() []string {
 	}
 }
 
-func (d *DevelopmentDockerComposeBuilder) getDockerVolumes() []string {
+func (d *ServiceComposeBuilder) getDockerVolumes() []string {
 	if !d.Mode.Mount {
 		return []string{}
 	}
 	return []string{d.getServiceFilePath() + ":" + "/mnt"}
 }
 
-func (d *DevelopmentDockerComposeBuilder) getRestartPolicy() string {
+func (d *ServiceComposeBuilder) getRestartPolicy() string {
 	if d.Mode.Environment != BuildModeEnvironmentTest {
 		return "on-failure"
 	}
 	return ""
 }
 
-func (d *DevelopmentDockerComposeBuilder) getInternalServiceDockerConfigs() (types.DockerConfigs, error) {
+func (d *ServiceComposeBuilder) getInternalServiceDockerConfigs() (types.DockerConfigs, error) {
 	result := types.DockerConfigs{}
 	result[d.Role] = types.DockerConfig{
 		Build: map[string]string{
@@ -115,7 +123,7 @@ func (d *DevelopmentDockerComposeBuilder) getInternalServiceDockerConfigs() (typ
 	return result.Merge(dependencyDockerConfigs), nil
 }
 
-func (d *DevelopmentDockerComposeBuilder) getExternalServiceDockerConfigs() (types.DockerConfigs, error) {
+func (d *ServiceComposeBuilder) getExternalServiceDockerConfigs() (types.DockerConfigs, error) {
 	result := types.DockerConfigs{}
 	if d.Mode.Environment == BuildModeEnvironmentTest {
 		return result, nil
@@ -132,11 +140,11 @@ func (d *DevelopmentDockerComposeBuilder) getExternalServiceDockerConfigs() (typ
 	return result, nil
 }
 
-func (d *DevelopmentDockerComposeBuilder) getServiceFilePath() string {
+func (d *ServiceComposeBuilder) getServiceFilePath() string {
 	return path.Join("${APP_PATH}", d.ServiceData.Location)
 }
 
-func (d *DevelopmentDockerComposeBuilder) getDockerEnvVars() map[string]string {
+func (d *ServiceComposeBuilder) getDockerEnvVars() map[string]string {
 	result := map[string]string{"ROLE": d.Role}
 	for _, builtDependency := range d.BuiltAppDependencies {
 		for variable, value := range builtDependency.GetServiceEnvVariables() {
@@ -148,7 +156,14 @@ func (d *DevelopmentDockerComposeBuilder) getDockerEnvVars() map[string]string {
 			result[variable] = value
 		}
 	}
-	envVars, secrets := d.ServiceConfig.GetEnvVars("development")
+	var envVars map[string]string
+	var secrets []string
+	switch d.Mode.Environment {
+	case BuildModeEnvironmentProduction:
+		envVars, secrets = d.ServiceConfig.GetEnvVars("production")
+	default:
+		envVars, secrets = d.ServiceConfig.GetEnvVars("development")
+	}
 	util.Merge(result, envVars)
 	for _, secret := range secrets {
 		result[secret] = os.Getenv(secret)
@@ -158,7 +173,7 @@ func (d *DevelopmentDockerComposeBuilder) getDockerEnvVars() map[string]string {
 	return result
 }
 
-func (d *DevelopmentDockerComposeBuilder) createServiceEndpointEnvVars() map[string]string {
+func (d *ServiceComposeBuilder) createServiceEndpointEnvVars() map[string]string {
 	endpoints := map[string]string{}
 	for role, serviceEndpoint := range d.ServiceEndpoints {
 		if role == d.Role {
@@ -169,7 +184,7 @@ func (d *DevelopmentDockerComposeBuilder) createServiceEndpointEnvVars() map[str
 	return endpoints
 }
 
-func (d *DevelopmentDockerComposeBuilder) getServiceDependsOn() []string {
+func (d *ServiceComposeBuilder) getServiceDependsOn() []string {
 	result := []string{}
 	for _, builtDependency := range d.BuiltAppDependencies {
 		result = append(result, builtDependency.GetContainerName())
@@ -182,7 +197,7 @@ func (d *DevelopmentDockerComposeBuilder) getServiceDependsOn() []string {
 }
 
 // returns the DockerConfigs object for a service's dependencies
-func (d *DevelopmentDockerComposeBuilder) getServiceDependenciesDockerConfigs() (types.DockerConfigs, error) {
+func (d *ServiceComposeBuilder) getServiceDependenciesDockerConfigs() (types.DockerConfigs, error) {
 	result := types.DockerConfigs{}
 	for _, builtDependency := range d.BuiltServiceDependencies {
 		dockerConfig, err := builtDependency.GetDockerConfig()
